requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
botbuilder-integration-aiohttp==4.16.1
litellm==1.67.0
chromadb==0.6.3
httpx==0.28.1
python-dotenv==1.0.1
redis==5.2.1
